actions: update to macos-14 because macos-13 will be removed in December 2025
We might as well do it for 8.3.0, so that it and 9.3.x patches can hopefully target the same version.

This means that our CI now runs on ARM64 instead of x86_64 for macOS.

To build HashLink, we need to install x86_64 Homebrew to get the necessary dependencies.

We also need a new enough krdlab/setup-haxe that supports ARM64 macOS, and can run older Haxe versions. This is currently commit 8f35d1215b93e940a76f9470e22e8a5ba6149598, but hopefully, setup-haxe's v2 tag will be updated in the future.